[lib] Refactor client into separate files
So it's easier to follow, file was getting to big.
Each "modifier" (throttle, retries...) in its own file
Simplify usage and avoid code repetition by introducing the executor
Client

go test ./...
go run redhat2nvd ...
